In progress: integrating pgvec.py into test_vector_stores

+++ langroid/vector_store/pgvec.py
@@ -26,7 +26,7 @@ from langroid.embedding_models.base import (
     EmbeddingModelsConfig,
 )
 from langroid.embedding_models.models import OpenAIEmbeddingsConfig
-from langroid.mytypes import Document
+from langroid.mytypes import Document, DocMetaData
 from langroid.utils.configuration import settings
 from langroid.utils.output.printing import print_long_text
 from langroid.vector_store.base import VectorStore, VectorStoreConfig
@@ -78,8 +78,6 @@ class PGVectorConfig(VectorStoreConfig):
     username: str = "postgres"
     password: str = "postgres"
     database: str = "langroid"
-    replace_collection: bool = False
-    embedding: EmbeddingModelsConfig = OpenAIEmbeddingsConfig()
     use_jsonb: bool = True
 
 def _get_embedding_table_class(table_name: str, vector_dimension: Optional[int] = None) -> Any:
@@ -117,9 +115,9 @@ def _create_vector_extension(conn: Engine) -> None:
 
 class PGVector(VectorStore):
     def __init__(self, config: PGVectorConfig = PGVectorConfig()):
-        super().__init__(config)
+        super().__init__(config) # reintroducing this is failing table creation
         self.config: PGVectorConfig = config
-        self.embedding_fn = self.embedding_model.embedding_fn()
+        self.embedding_fn = self.embedding_model.embedding_fn() # embedding_model set in super().__init__
         self.embedding_dim = self.embedding_model.embedding_dims
 
         if Vector is None:
@@ -179,8 +177,6 @@ class PGVector(VectorStore):
         }
 
     def set_collection(self, collection_name: str, replace: bool = False) -> None:
-        # if collection_name == self.config.collection_name:
-        #     return
         if collection_name in Base.metadata.tables:
             return
         
@@ -259,8 +255,13 @@ class PGVector(VectorStore):
     def _generate_ids(documents: Sequence[Document]) -> List[str]:
         return [str(uuid.uuid5(uuid.NAMESPACE_DNS, d.content)) for d in documents]
 
+    def maybe_add_ids(self, documents: Sequence[Document]) -> None:
+        for d in documents:
+            if d.metadata.id is None or d.metadata.id == "":
+                d.metadata.id = str(uuid.uuid4())
+
     def add_documents(self, documents: Sequence[Document]) -> None:
-        super().maybe_add_ids(documents)
+        self.maybe_add_ids(documents)
         with self.SessionLocal() as session:
             embeddings = [self.embedding_fn([d.content])[0] for d in documents]
             ids = self._generate_ids(documents)
@@ -286,11 +287,13 @@ class PGVector(VectorStore):
     def delete_collection(self, collection_name: str) -> None:
         if collection_name in _classes: # clear cached embedding table
             EmbeddingTable = _classes.pop(collection_name)
-            
+            self.metadata.clear() # Clear metadata cache
             metadata = MetaData()
-            metadata.reflect(bind=self.engine) # get the current db state
-            table = metadata.tables[collection_name]
-            table.drop(self.engine, checkfirst=True) # drop table and update db
+            metadata.reflect(bind=self.engine)
+            
+            if collection_name in metadata.tables:
+                table = metadata.tables[collection_name]
+                table.drop(self.engine, checkfirst=True) # drop table and update db
            
         else:
             logger.warning(f"Table '{collection_name}' not found.")
@@ -340,166 +343,169 @@ def test_create_collection():
 
     postgres_db.create_collection("test_collection1", replace=True)
 
+    postgres_db.refresh_metadata()
     inspector = inspect(engine)
-    assert inspector.has_table("test_collection1"), "Table 'test_collection' does not exist"
+    assert inspector.has_table("test_collection1"), "Table 'test_collection1' does not exist"
     
     postgres_db.delete_collection("test_collection1")
+
+    postgres_db.refresh_metadata()
     inspector = inspect(engine)
-    assert not inspector.has_table("test_collection1"), "Table 'test_collection' was not deleted"
 
-# def test_list_collections():
-#     db_config = PGVectorConfig(
-#         username="postgres",
-#         embedding=embed_cfg,
-#     )
-#     try:
-#         postgres_db = PGVector(db_config)
-#     except Exception as e:
-#         print(f"Error initializing PGVector: {e}")
-#         return
+    assert not inspector.has_table("test_collection1"), "Table 'test_collection1' was not deleted"
 
-#     postgres_db.create_collection("test_collection_1")
-#     postgres_db.create_collection("test_collection_2")
-#     postgres_db.create_collection("test_collection_3")
+def test_list_collections():
+    db_config = PGVectorConfig(
+        username="postgres",
+    )
+    try:
+        postgres_db = PGVector(db_config)
+    except Exception as e:
+        print(f"Error initializing PostgresDB: {e}")
+        return
 
-#     collections = postgres_db.list_collections()
-#     print(collections)
-#     assert "test_collection_1" in collections, "Collection 'test_collection_1' not found"
-#     assert "test_collection_2" in collections, "Collection 'test_collection_2' not found"
-#     assert "test_collection_3" in collections, "Collection 'test_collection_3' not found"
+    postgres_db.create_collection("test_collection_1")
+    postgres_db.create_collection("test_collection_2")
+    postgres_db.create_collection("test_collection_3")
 
-#     postgres_db.delete_collection("test_collection_1")
-#     postgres_db.delete_collection("test_collection_2")
-#     postgres_db.delete_collection("test_collection_3")
+    collections = postgres_db.list_collections()
+    print(collections)
+    assert "test_collection_1" in collections, "Collection 'test_collection_1' not found"
+    assert "test_collection_2" in collections, "Collection 'test_collection_2' not found"
+    assert "test_collection_3" in collections, "Collection 'test_collection_3' not found"
 
-#     collections = postgres_db.list_collections()
-#     print(collections)
-#     assert "test_collection_1" not in collections, "Collection 'test_collection_1' was not deleted"
-#     assert "test_collection_2" not in collections, "Collection 'test_collection_2' was not deleted"
-#     assert "test_collection_3" not in collections, "Collection 'test_collection_3' was not deleted"
+    postgres_db.delete_collection("test_collection_1")
+    postgres_db.delete_collection("test_collection_2")
+    postgres_db.delete_collection("test_collection_3")
 
-# def test_add_documents():
-#     db_config = PGVectorConfig(
-#         collection_name="test_add_docs_collection",
-#         username="postgres",
-#         embedding=embed_cfg
-#     )
-#     try:
-#         postgres_db = PGVector(db_config)
-#     except Exception as e:
-#         print(f"Error initializing PGVector: {e}")
-#         return
+    collections = postgres_db.list_collections()
+    print(collections)
+    assert "test_collection_1" not in collections, "Collection 'test_collection_1' was not deleted"
+    assert "test_collection_2" not in collections, "Collection 'test_collection_2' was not deleted"
+    assert "test_collection_3" not in collections, "Collection 'test_collection_3' was not deleted"
 
-#     postgres_db.create_collection("test_add_docs_collection")
+def test_add_documents():
+    db_config = PGVectorConfig(
+        collection_name="test_add_docs_collection",
+        username="postgres",
+        embedding=embed_cfg
+    )
+    try:
+        postgres_db = PGVector(db_config)
+    except Exception as e:
+        print(f"Error initializing PGVector: {e}")
+        return
 
-#     doc1 = Document(content="document 1", metadata={"id": "1", "info": "info1"})
-#     doc2 = Document(content="document 2", metadata={"id": "2", "info": "info2"})
+    postgres_db.create_collection("test_add_docs_collection")
 
-#     postgres_db.add_documents([doc1, doc2])
+    doc1 = Document(content="document 1", metadata={"id": "1", "info": "info1"})
+    doc2 = Document(content="document 2", metadata={"id": "2", "info": "info2"})
 
-#     docs = postgres_db.get_all_documents()
-#     print(docs)
-#     assert len(docs) == 2, "Incorrect number of documents added"
-#     assert any(doc.content == "document 1" for doc in docs), "Document 1 not found"
-#     assert any(doc.content == "document 2" for doc in docs), "Document 2 not found"
+    postgres_db.add_documents([doc1, doc2])
 
-#     postgres_db.delete_collection("test_add_docs_collection")
+    docs = postgres_db.get_all_documents()
+    print(docs)
+    assert len(docs) == 2, "Incorrect number of documents added"
+    assert any(doc.content == "document 1" for doc in docs), "Document 1 not found"
+    assert any(doc.content == "document 2" for doc in docs), "Document 2 not found"
 
-# def test_get_documents_by_ids():
-#     db_config = PGVectorConfig(
-#         collection_name="test_get_docs_collection",
-#         username="postgres",
-#         embedding=embed_cfg
-#     )
-#     try:
-#         postgres_db = PGVector(db_config)
-#     except Exception as e:
-#         print(f"Error initializing PGVector: {e}")
-#         return
+    postgres_db.delete_collection("test_add_docs_collection")
 
-#     postgres_db.create_collection("test_get_docs_collection")
+def test_get_documents_by_ids():
+    db_config = PGVectorConfig(
+        collection_name="test_get_docs_collection",
+        username="postgres",
+        embedding=embed_cfg
+    )
+    try:
+        postgres_db = PGVector(db_config)
+    except Exception as e:
+        print(f"Error initializing PGVector: {e}")
+        return
 
-#     doc1 = Document(content="document 1", metadata={"id": "id-1"})
-#     doc2 = Document(content="document 2", metadata={"id": "id-2"})
-#     doc3 = Document(content="document 3", metadata={"id": "id-3"})
+    postgres_db.create_collection("test_get_docs_collection")
 
-#     postgres_db.add_documents([doc1, doc2, doc3])
+    doc1 = Document(content="document 1", metadata={"id": "id-1"})
+    doc2 = Document(content="document 2", metadata={"id": "id-2"})
+    doc3 = Document(content="document 3", metadata={"id": "id-3"})
 
-#     id_1 = str(uuid.uuid5(uuid.NAMESPACE_DNS, doc1.content))
-#     id_3 = str(uuid.uuid5(uuid.NAMESPACE_DNS, doc3.content))
+    postgres_db.add_documents([doc1, doc2, doc3])
 
-#     retrieved_docs = postgres_db.get_documents_by_ids([id_1, id_3])
-#     print(retrieved_docs)
-#     assert len(retrieved_docs) == 2, "Incorrect number of documents retrieved"
-#     assert any(doc.content == "document 1" for doc in retrieved_docs), "Document 1 not found"
-#     assert any(doc.content == "document 3" for doc in retrieved_docs), "Document 3 not found"
+    id_1 = str(uuid.uuid5(uuid.NAMESPACE_DNS, doc1.content))
+    id_3 = str(uuid.uuid5(uuid.NAMESPACE_DNS, doc3.content))
 
-#     postgres_db.delete_collection("test_get_docs_collection")
+    retrieved_docs = postgres_db.get_documents_by_ids([id_1, id_3])
+    print(retrieved_docs)
+    assert len(retrieved_docs) == 2, "Incorrect number of documents retrieved"
+    assert any(doc.content == "document 1" for doc in retrieved_docs), "Document 1 not found"
+    assert any(doc.content == "document 3" for doc in retrieved_docs), "Document 3 not found"
 
-# def test_get_all_documents():
-#     db_config = PGVectorConfig(
-#         collection_name="test_get_all_docs_collection",
-#         username="postgres",
-#         embedding=embed_cfg
-#     )
-#     try:
-#         postgres_db = PGVector(db_config)
-#     except Exception as e:
-#         print(f"Error initializing PGVector: {e}")
-#         return
+    postgres_db.delete_collection("test_get_docs_collection")
 
-#     postgres_db.create_collection("test_get_all_docs_collection")
+def test_get_all_documents():
+    db_config = PGVectorConfig(
+        collection_name="test_get_all_docs_collection",
+        username="postgres",
+        embedding=embed_cfg
+    )
+    try:
+        postgres_db = PGVector(db_config)
+    except Exception as e:
+        print(f"Error initializing PGVector: {e}")
+        return
 
-#     doc1 = Document(content="document 1", metadata={"id": "1"})
-#     doc2 = Document(content="document 2", metadata={"id": "2"})
-#     doc3 = Document(content="document 3", metadata={"id": "3"})
+    postgres_db.create_collection("test_get_all_docs_collection")
 
-#     postgres_db.add_documents([doc1, doc2, doc3])
+    doc1 = Document(content="document 1", metadata={"id": "1"})
+    doc2 = Document(content="document 2", metadata={"id": "2"})
+    doc3 = Document(content="document 3", metadata={"id": "3"})
 
-#     all_docs = postgres_db.get_all_documents()
-#     print(all_docs)
-#     assert len(all_docs) == 3, "Incorrect number of documents retrieved"
-#     assert any(doc.content == "document 1" for doc in all_docs), "Document 1 not found"
-#     assert any(doc.content == "document 2" for doc in all_docs), "Document 2 not found"
-#     assert any(doc.content == "document 3" for doc in all_docs), "Document 3 not found"
+    postgres_db.add_documents([doc1, doc2, doc3])
 
-#     postgres_db.delete_collection("test_get_all_docs_collection")
+    all_docs = postgres_db.get_all_documents()
+    print(all_docs)
+    assert len(all_docs) == 3, "Incorrect number of documents retrieved"
+    assert any(doc.content == "document 1" for doc in all_docs), "Document 1 not found"
+    assert any(doc.content == "document 2" for doc in all_docs), "Document 2 not found"
+    assert any(doc.content == "document 3" for doc in all_docs), "Document 3 not found"
 
-# def test_clear_all_collections():
-#     db_config = PGVectorConfig(
-#         username="postgres",
-#         embedding=embed_cfg
-#     )
-#     try:
-#         postgres_db = PGVector(db_config)
-#     except Exception as e:
-#         print(f"Error initializing PGVector: {e}")
-#         return
+    postgres_db.delete_collection("test_get_all_docs_collection")
 
-#     postgres_db.create_collection("test_collection_1")
-#     postgres_db.create_collection("test_collection_2")
-#     postgres_db.create_collection("test_collection_3")
+def test_clear_all_collections():
+    db_config = PGVectorConfig(
+        username="postgres",
+        embedding=embed_cfg
+    )
+    try:
+        postgres_db = PGVector(db_config)
+    except Exception as e:
+        print(f"Error initializing PGVector: {e}")
+        return
 
-#     doc1 = Document(content="doc1 in collection 1", metadata={"id": "d1"})
-#     doc2 = Document(content="doc2 in collection 2", metadata={"id": "d2"})
-#     doc3 = Document(content="doc3 in collection 3", metadata={"id": "d3"})
-#     postgres_db.add_documents([doc1])
-#     postgres_db.set_collection("test_collection_2")
-#     postgres_db.add_documents([doc2])
-#     postgres_db.set_collection("test_collection_3")
-#     postgres_db.add_documents([doc3])
-#     postgres_db.set_collection("test_collection_1")
+    postgres_db.create_collection("test_collection_1")
+    postgres_db.create_collection("test_collection_2")
+    postgres_db.create_collection("test_collection_3")
 
-#     num_deleted = postgres_db.clear_all_collections(really=True)
-#     assert num_deleted == 3, "Incorrect number of collections deleted"
+    doc1 = Document(content="doc1 in collection 1", metadata={"id": "d1"})
+    doc2 = Document(content="doc2 in collection 2", metadata={"id": "d2"})
+    doc3 = Document(content="doc3 in collection 3", metadata={"id": "d3"})
+    postgres_db.add_documents([doc1])
+    postgres_db.set_collection("test_collection_2")
+    postgres_db.add_documents([doc2])
+    postgres_db.set_collection("test_collection_3")
+    postgres_db.add_documents([doc3])
+    postgres_db.set_collection("test_collection_1")
 
-#     collections = postgres_db.list_collections()
-#     assert len(collections) == 0, "Collections were not cleared"
+    num_deleted = postgres_db.clear_all_collections(really=True)
+    assert num_deleted == 3, "Incorrect number of collections deleted"
+
+    collections = postgres_db.list_collections()
+    assert len(collections) == 0, "Collections were not cleared"
 
 # # Run the test functions
-test_create_collection()
-# # test_list_collections()
-# # test_add_documents()
-# # test_get_documents_by_ids()
-# # test_get_all_documents()
-# # test_clear_all_collections()
+# test_create_collection()
+# test_list_collections()
+# test_add_documents()
+# test_get_documents_by_ids()
+# test_get_all_documents()
+# test_clear_all_collections()
